피드백 반영 2
- 중복되는 테스트 함수 호출로 변경: fixtures폴더&restaurants.js 생성
- 파일 이름변경
