+++ fixtures/restaurants.js
@@ -1,0 +1,10 @@
+const restaurants = [
+  {
+    id: 1,
+    name: '마녀주방',
+    category: '한식',
+    address: '서울시 강남구',
+  },
+];
+
+export default restaurants;
