--- src/List.jsx
@@ -1,11 +1,0 @@
-export default function List({ restaurants }) {
-  return (
-    <ul>
-      {restaurants.map((restaurant) => (
-        <li key={restaurant.id}>
-          {`${restaurant.name} | ${restaurant.category} | ${restaurant.address}`}
-        </li>
-      ))}
-    </ul>
-  );
-}
